Address PR feedback

+++ src/System.Memory/src/System/NUint.cs
@@ -44,19 +44,19 @@ namespace System
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static NUInt operator *(NUInt left, NUInt right)
         {
-            unsafe { return (sizeof(IntPtr) == 4) ? new NUInt(((uint)left._value) * (uint)right._value) : new NUInt(((ulong)left._value) * (uint)right._value); }
+            unsafe { return (sizeof(IntPtr) == 4) ? new NUInt(((uint)left._value) * (uint)right._value) : new NUInt(((ulong)left._value) * (ulong)right._value); }
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static NUInt operator +(NUInt left, NUInt right)
         {
-            unsafe { return (sizeof(IntPtr) == 4) ? new NUInt(((uint)left._value) + (uint)right._value) : new NUInt(((ulong)left._value) + (uint)right._value); }
+            unsafe { return (sizeof(IntPtr) == 4) ? new NUInt(((uint)left._value) + (uint)right._value) : new NUInt(((ulong)left._value) + (ulong)right._value); }
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static NUInt operator -(NUInt left, NUInt right)
         {
-            unsafe { return (sizeof(IntPtr) == 4) ? new NUInt(((uint)left._value) - (uint)right._value) : new NUInt(((ulong)left._value) - (uint)right._value); }
+            unsafe { return (sizeof(IntPtr) == 4) ? new NUInt(((uint)left._value) - (uint)right._value) : new NUInt(((ulong)left._value) - (ulong)right._value); }
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -78,12 +78,12 @@ namespace System
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool operator <=(NUInt left, NUInt right)
         {
-            unsafe { return (sizeof(IntPtr) == 4) ? ((uint)left._value) <= (uint)right._value : ((ulong)left._value) <= (uint)right._value; }
+            unsafe { return (sizeof(IntPtr) == 4) ? ((uint)left._value) <= (uint)right._value : ((ulong)left._value) <= (ulong)right._value; }
         }
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool operator >=(NUInt left, NUInt right)
         {
-            unsafe { return (sizeof(IntPtr) == 4) ? ((uint)left._value) >= (uint)right._value : ((ulong)left._value) >= (uint)right._value; }
+            unsafe { return (sizeof(IntPtr) == 4) ? ((uint)left._value) >= (uint)right._value : ((ulong)left._value) >= (ulong)right._value; }
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
